refactor(database): move RevisionRange from metadata to offline
